apps/status-api/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pyzmq==26.4.0
prometheus-client==0.22.1
